split association overrides to individual files
