+++ lib/hoardable.rb
@@ -10,5 +10,9 @@ require_relative 'hoardable/source_model'
 require_relative 'hoardable/version_model'
 require_relative 'hoardable/model'
 require_relative 'hoardable/associations'
+require_relative 'hoardable/has_many'
+require_relative 'hoardable/belongs_to'
+require_relative 'hoardable/has_one'
+require_relative 'hoardable/has_rich_text'
 require_relative 'generators/hoardable/migration_generator'
 require_relative 'generators/hoardable/install_generator'
